Extract query class for NotificationDefault.
